+++ CHANGELOG.md
@@ -4,6 +4,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 ## Unreleased
 
+- Support for loading JQ [modules](https://stedolan.github.io/jq/manual/#Modules)
 - Fixed a bug for musl based release #24
 - Updated `net.thisptr/jackson-jq` to `1.0.0-preview.20220705`
 - Updated GraalVM to 22.2.0